+++ TagScriptEngine/block/breakblock.py
@@ -1,16 +1,16 @@
 from typing import Optional
 
-from .. import Interpreter, adapter
 from ..interface import Block
+from ..interpreter import Context
 from . import helper_parse_if
 
 
 class BreakBlock(Block):
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return any([dec == "break", dec == "shortcircuit", dec == "short"])
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if helper_parse_if(ctx.verb.parameter) == True:
             ctx.response.body = ctx.verb.payload if ctx.verb.payload != None else ""
         return ""
