+++ TagScriptEngine/block/fiftyfifty.py
@@ -1,16 +1,16 @@
 import random
 from typing import Optional
 
-from .. import Interpreter, adapter
 from ..interface import Block
+from ..interpreter import Context
 
 
 class FiftyFiftyBlock(Block):
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return any([dec == "5050", dec == "50", dec == "?"])
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if ctx.verb.payload is None:
             return None
         return random.choice(["", ctx.verb.payload])
