+++ TagScriptEngine/block/randomblock.py
@@ -1,8 +1,8 @@
 import random
 from typing import Optional
 
-from .. import Interpreter, adapter
 from ..interface import Block
+from ..interpreter import Context
 
 
 class RandomBlock(Block):
@@ -33,11 +33,11 @@ class RandomBlock(Block):
         # Assigns a random insult to the insult variable
     """
 
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return dec in ("random", "#", "rand")
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if ctx.verb.payload is None:
             return None
         spl = []
