+++ TagScriptEngine/block/require_blacklist.py
@@ -1,30 +1,7 @@
-"""
-MIT License
-
-Copyright (c) 2020-2021 phenom4n4n
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-"""
-
 from typing import Optional
 
-from TagScriptEngine import Block, Interpreter
+from ..interface import Block
+from ..interpreter import Context
 
 
 class RequireBlock(Block):
@@ -50,11 +27,11 @@ class RequireBlock(Block):
         {require(757425366209134764, 668713062186090506, 737961895356792882):You aren't allowed to use this tag.}
     """
 
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return any([dec == "require", dec == "whitelist"])
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if not ctx.verb.parameter:
             return None
         actions = ctx.response.actions.get("requires")
@@ -88,11 +65,11 @@ class BlacklistBlock(Block):
         {blacklist(Tag Blacklist, 668713062186090506):You are blacklisted from using tags.}
     """
 
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return dec == "blacklist"
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if not ctx.verb.parameter:
             return None
         actions = ctx.response.actions.get("blacklist")
