+++ TagScriptEngine/block/shortcutredirect.py
@@ -1,17 +1,18 @@
 from typing import Optional
 
-from .. import Interpreter, Verb, adapter
 from ..interface import Block
+from ..interpreter import Context
+from ..verb import Verb
 
 
 class ShortCutRedirectBlock(Block):
     def __init__(self, var_name):
         self.redirect_name = var_name
 
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         return ctx.verb.declaration.isdigit()
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         blank = Verb()
         blank.declaration = self.redirect_name
         blank.parameter = ctx.verb.declaration
