+++ TagScriptEngine/block/stopblock.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from .. import Interpreter, adapter
 from ..interface import Block
+from ..interpreter import Context
 from . import helper_parse_if
 
 
@@ -24,11 +24,11 @@ class StopBlock(Block):
         # enforces providing arguments for a tag
     """
 
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         dec = ctx.verb.declaration.lower()
         return dec in ("stop", "halt", "error")
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         if ctx.verb.parameter is None:
             return None
         if helper_parse_if(ctx.verb.parameter):
