+++ TagScriptEngine/block/strictvariablegetter.py
@@ -1,12 +1,12 @@
 from typing import Optional
 
-from .. import Interpreter, adapter
 from ..interface import Block
+from ..interpreter import Context
 
 
 class StrictVariableGetterBlock(Block):
-    def will_accept(self, ctx: Interpreter.Context) -> bool:
+    def will_accept(self, ctx: Context) -> bool:
         return ctx.verb.declaration in ctx.response.variables
 
-    def process(self, ctx: Interpreter.Context) -> Optional[str]:
+    def process(self, ctx: Context) -> Optional[str]:
         return ctx.response.variables[ctx.verb.declaration].get_value(ctx.verb)
