+++ TagScriptEngine/interface/adapter.py
@@ -1,12 +1,14 @@
-from .. import Verb
+from typing import Optional
+
+from ..interpreter import Context
 
 
-class Adapter(object):
+class Adapter:
     def __init__(self):
         pass
 
     def __repr__(self):
         return f"<{type(self).__qualname__} at {hex(id(self))}>"
 
-    def get_value(self, ctx: Verb) -> str:
+    def get_value(self, ctx: Context) -> Optional[str]:
         return ""
